requirements.txt
pytelegrambotapi
python-twitter